# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Innovatech Solutions",
    "version": "1.0.0",
    "description": "Documentación de los Microservicios Serverless de Gestión."
  },
  "paths": {
...
    "/api/proyectos/{id}": {
      "delete": {
        "summary": "Eliminar un proyecto específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID único del proyecto a eliminar"
          }
        ],
        "responses": {
          "200": { "description": "Proyecto eliminado exitosamente" }
        }
      }
    },
...
  },
...
}